Extract reusable catalog product card
